requirements.txt
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchain-chroma==0.1.1
chromadb==0.5.0
python-dotenv==1.0.1
requests==2.31.0
unstructured==0.14.0
markdown==3.6
networkx==3.1
gradio==4.36.1